backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
Pillow==10.2.0
